Warnings.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@45587 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/utils.cpp
@@ -356,7 +356,7 @@ bool prefix_kernel_object_name(char* name, size_t bufsize)
 
 		// if name and prefix can't fit in name's buffer than we must
 		// not overwrite end of name because it contains object type
-		const int move_prefix = (len_name + len_prefix > bufsize) ?
+		const size_t move_prefix = (len_name + len_prefix > bufsize) ?
 			(bufsize - len_name) : len_prefix;
 
 		memmove(name + move_prefix, name, len_name);
